Rename timeline folder in patientview
