+++ src/pages/patientView/PatientViewPage.tsx
@@ -74,10 +74,10 @@ import ResourcesTab, { RESOURCES_TAB_NAME } from './resources/ResourcesTab';
 import { MakeMobxView } from '../../shared/components/MobxView';
 import ResourceTab from '../../shared/components/resources/ResourceTab';
 import PatientViewStructuralVariantTable from './structuralVariant/PatientViewStructuralVariantTable';
-import TimelineWrapper from './timeline2/TimelineWrapper';
+import TimelineWrapper from './timeline/TimelineWrapper';
 import { isFusion } from '../../shared/lib/MutationUtils';
 import { Mutation } from 'cbioportal-ts-api-client';
-import ClinicalEventsTables from './timeline2/ClinicalEventsTables';
+import ClinicalEventsTables from './timeline/ClinicalEventsTables';
 import MutationalSignaturesContainer from './mutationalSignatures/MutationalSignaturesContainer';
 import SampleSummaryList from './sampleHeader/SampleSummaryList';
 import { updateOncoKbIconStyle } from 'shared/lib/AnnotationColumnUtils';
