+++ src/pages/patientView/timeline/ClinicalEventsTables.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { ClinicalEvent } from 'cbioportal-ts-api-client';
-import { groupTimelineData } from 'pages/patientView/timeline2/timelineDataUtils.ts';
+import { groupTimelineData } from 'pages/patientView/timeline/timelineDataUtils.ts';
 import LazyMobXTable from 'shared/components/lazyMobXTable/LazyMobXTable';
 import _ from 'lodash';
