+++ src/pages/patientView/timeline/styles.scss
